Rename 'iso_3' to 'iso3' etc.
